Move company export code to modules
