+++ src/client/modules/Companies/CompanyExports/ExportHistory/reducer.js
@@ -1,7 +1,7 @@
 import {
   EXPORTS_HISTORY__LOADED,
   EXPORTS_HISTORY__SELECT_PAGE,
-} from '../../../../../../client/actions'
+} from '../../../../actions'
 
 const initialState = {
   count: 0,
