+++ src/client/modules/Companies/CompanyExports/ExportWins/index.jsx
@@ -3,12 +3,12 @@ import styled from 'styled-components'
 import { SPACING } from '@govuk-react/constants'
 import { connect } from 'react-redux'
 
-import { CollectionList } from '../../../../../../client/components/'
+import { CollectionList } from '../../../../components'
 import { state2props, NOT_IMPLEMENTED } from './state'
 import {
   EXPORT_WINS__LOADED,
   EXPORT_WINS__SELECT_PAGE,
-} from '../../../../../../client/actions'
+} from '../../../../actions'
 
 const Wrapper = styled('div')`
   margin-top: ${SPACING.SCALE_3};
